Use new AgentRunJobHistory api.
The AgentRunJobHistory has been updated to match the latest changes in
GoCD 20.3.0. The main differences is that the data structures for the
agent job history has changed along with the timestamp for job state
transition.

The newer API also support pagination, but it also added a page_size
parameter. We now use the page_size parameter and set it to its maximum
allowed value of 100. This means that we don't need to paginate any
further as 100 last jobs should be good enough always. All pagination
references has been removed completely.

The jobs listed in agent job history does not longer have an ID
associated so we're creating our own unique ID. This is mainly for
caching so that we don't add the same job twice as prometheus metrics.

As the timestamps in AgentJobStateTransition has changed we need to
convert them to unix timestamps so that we can compare them easier. This
could probably be done in a better way using the time package directly
for comparing but the current implementation works good enough.